move diffusionMatrix2D to stochastic_process
